+++ app/javascript/components/index.js
@@ -10,3 +10,4 @@ export * from './TransfersStatus';
 export * from './Week';
 export * from './FantasyLeague';
 export * from './FeedbackForm';
+export * from './FantasyTeamDestroyForm';